move spine framework files to spm package
