--- spine-ios/Example/Spine iOS Example/AAPLShaders.metal
@@ -1,110 +1,0 @@
-/*
-See LICENSE folder for this sample’s licensing information.
-
-Abstract:
-Metal shaders used for this sample
-*/
-
-#include <metal_stdlib>
-
-using namespace metal;
-
-// TODO: Move to spine package and import correct header
-// Include header shared between this Metal shader code and C code executing Metal API commands.
-//#include "AAPLShaderTypes.h"
-
-// Buffer index values shared between shader and C code to ensure Metal shader buffer inputs
-// match Metal API buffer set calls.
-typedef enum AAPLVertexInputIndex
-{
-    AAPLVertexInputIndexVertices     = 0,
-    AAPLVertexInputIndexTransform    = 1,
-    AAPLVertexInputIndexViewportSize = 2,
-} AAPLVertexInputIndex;
-
-// Texture index values shared between shader and C code to ensure Metal shader buffer inputs match
-//   Metal API texture set calls
-typedef enum AAPLTextureIndex
-{
-    AAPLTextureIndexBaseColor = 0,
-} AAPLTextureIndex;
-
-//  This structure defines the layout of vertices sent to the vertex
-//  shader. This header is shared between the .metal shader and C code, to guarantee that
-//  the layout of the vertex array in the C code matches the layout that the .metal
-//  vertex shader expects.
-typedef struct
-{
-    vector_float2 position;
-    vector_float4 color;
-    vector_float2 uv;
-} AAPLVertex;
-
-typedef struct {
-    vector_float2 translation;
-    vector_float2 scale;
-    vector_float2 offset;
-} AAPLTransform;
-
-// Vertex shader outputs and fragment shader inputs
-struct RasterizerData
-{
-    // The [[position]] attribute of this member indicates that this value
-    // is the clip space position of the vertex when this structure is
-    // returned from the vertex function.
-    float4 position [[position]];
-
-    // Since this member does not have a special attribute, the rasterizer
-    // interpolates its value with the values of the other triangle vertices
-    // and then passes the interpolated value to the fragment shader for each
-    // fragment in the triangle.
-    float4 color;
-    
-    float2 textureCoordinate;
-};
-
-vertex RasterizerData
-vertexShader(uint vertexID [[vertex_id]],
-             constant AAPLVertex *vertices [[buffer(AAPLVertexInputIndexVertices)]],
-             constant AAPLTransform *transform [[buffer(AAPLVertexInputIndexTransform)]],
-             constant vector_uint2 *viewportSizePointer [[buffer(AAPLVertexInputIndexViewportSize)]])
-{
-    RasterizerData out;
-
-    // Index into the array of positions to get the current vertex.
-    // The positions are specified in pixel dimensions (i.e. a value of 100
-    // is 100 pixels from the origin).
-    float2 pixelSpacePosition = vertices[vertexID].position.xy;
-
-    // Get the viewport size and cast to float.
-    vector_float2 viewportSize = vector_float2(*viewportSizePointer);
-    
-    out.position = vector_float4(0.0, 0.0, 0.0, 1.0);
-    
-    out.position.xy = pixelSpacePosition;
-    out.position.xy *= transform->scale;
-    out.position.xy += transform->translation * transform->scale + transform->offset;
-    out.position.xy /= viewportSize / 2;
-    out.position.y *= -1;
-    
-    // Pass the input color directly to the rasterizer.
-    out.color = vertices[vertexID].color;
-    
-    out.textureCoordinate = vertices[vertexID].uv;
-    
-    return out;
-}
-
-fragment float4
-fragmentShader(RasterizerData in [[stage_in]],
-               texture2d<half> colorTexture [[ texture(AAPLTextureIndexBaseColor) ]])
-{
-    constexpr sampler textureSampler (mag_filter::nearest,
-                                      min_filter::nearest);
-    
-    // Sample the texture to obtain a color
-    const half4 colorSample = colorTexture.sample(textureSampler, in.textureCoordinate);
-    
-    return float4(colorSample) * in.color;
-}
-
