--- spine-ios/Example/Spine iOS Example/spine-ios/README.md
@@ -1,3 +1,0 @@
-# README
-
-Due to Xcode limitations, classes that are supposed to go to `spine-ios` package will live here for now.
