+++ spine-ios/Sources/Spine/SpineController.swift
@@ -1,12 +1,4 @@
-//
-//  ConentViewModel.swift
-//  Spine iOS Example
-//
-//  Created by Denis Andrašec on 08.05.24.
-//
-
 import Foundation
-import Spine
 import CoreGraphics
 import QuartzCore
 import UIKit
